Conflicts:
	src/openms_gui/source/VISUAL/APPLICATIONS/GUITOOLS/TOPPView.cpp

+++ src/openms/source/METADATA/SourceFile.cpp
@@ -2,7 +2,7 @@
 //                   OpenMS -- Open-Source Mass Spectrometry
 // --------------------------------------------------------------------------
 // Copyright The OpenMS Team -- Eberhard Karls University Tuebingen,
-// ETH Zurich, and Freie Universitaet Berlin 2002-2015.
+// ETH Zurich, and Freie Universitaet Berlin 2002-2017.
 //
 // This software is released under a three-clause BSD license:
 //  * Redistributions of source code must retain the above copyright
@@ -28,13 +28,11 @@
 // ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //
 // --------------------------------------------------------------------------
-// $Maintainer: Stephan Aiche $
+// $Maintainer: Timo Sachsenberg $
 // $Authors: Marc Sturm $
 // --------------------------------------------------------------------------
 
 #include <OpenMS/METADATA/SourceFile.h>
-
-#include <iostream>
 
 using namespace std;
 
@@ -132,7 +130,7 @@ namespace OpenMS
 
   void SourceFile::setFileSize(float file_size)
   {
-    file_size_ = file_size;
+    file_size_ = static_cast<double>(file_size);
   }
 
   const String& SourceFile::getFileType() const
